requirements.txt
fastapi
uvicorn
requests
pandas
python-dotenv
openpyxl
